+++ src/filterConfig.js
@@ -61,13 +61,13 @@ export const instanceFilterConfig = [
     name: FACETS.CREATED_DATE,
     cql: FACETS_CQL.CREATED_DATE,
     values: [],
-    parse: buildDateRangeQuery(FACETS.CREATED_DATE),
+    parse: buildDateRangeQuery(FACETS_CQL.CREATED_DATE),
   },
   {
     name: FACETS.UPDATED_DATE,
     cql: FACETS_CQL.UPDATED_DATE,
     values: [],
-    parse: buildDateRangeQuery(FACETS.UPDATED_DATE),
+    parse: buildDateRangeQuery(FACETS_CQL.UPDATED_DATE),
   },
   {
     name: FACETS.SOURCE,
@@ -140,6 +140,18 @@ export const holdingFilterConfig = [
     cql: FACETS_CQL.HOLDINGS_TAGS,
     values: [],
   },
+  {
+    name: FACETS.HOLDINGS_CREATED_DATE,
+    cql: FACETS_CQL.HOLDINGS_CREATED_DATE,
+    values: [],
+    parse: buildDateRangeQuery(FACETS_CQL.HOLDINGS_CREATED_DATE),
+  },
+  {
+    name: FACETS.HOLDINGS_UPDATED_DATE,
+    cql: FACETS_CQL.HOLDINGS_UPDATED_DATE,
+    values: [],
+    parse: buildDateRangeQuery(FACETS_CQL.HOLDINGS_UPDATED_DATE),
+  },
 ];
 
 export const itemIndexes = [
@@ -181,6 +193,18 @@ export const itemFilterConfig = [
     name: FACETS.ITEMS_DISCOVERY_SUPPRESS,
     cql: FACETS_CQL.ITEMS_DISCOVERY_SUPPRESS,
     values: [],
+  },
+  {
+    name: FACETS.ITEMS_CREATED_DATE,
+    cql: FACETS_CQL.ITEMS_CREATED_DATE,
+    values: [],
+    parse: buildDateRangeQuery(FACETS_CQL.ITEMS_CREATED_DATE),
+  },
+  {
+    name: FACETS.ITEMS_UPDATED_DATE,
+    cql: FACETS_CQL.ITEMS_UPDATED_DATE,
+    values: [],
+    parse: buildDateRangeQuery(FACETS_CQL.ITEMS_UPDATED_DATE),
   },
   {
     name: FACETS.ITEMS_TAGS,