+++ CHANGELOG.md
@@ -40,6 +40,10 @@
 * Add new data element `<AdministrativeNote>` to items. Refs UIIN-1444.
 * Browse form. Refs UIIN-1887.
 * Add Browse subjects option. Refs UIIN-1880.
+* Add ability to filter holding records by created date. Refs UIIN-791.
+* Add ability to filter holding records by updated date. Refs UIIN-786.
+* Add ability to filter item records by created date. Refs UIIN-789.
+* Add ability to filter item records by updated date. Refs UIIN-786.
 
 ## [8.0.0](https://github.com/folio-org/ui-inventory/tree/v8.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.4...v8.0.0)